Started re-creating the OneRepMax C# library (the original source was lost). The current NuGet package came to the rescue but I was not happy with the organization.
Only the functionality core to the concept of calculating a one rep maximum have been included. The strength comparison functionality has been excluded and will probably be put into a separate project. A new NuGet package and version will be assigned with a notification that this "update" will be (yet another) breaking change release.

No unit tests or demo have been created for this initial check-in.