Removed the Math.Round operation in the calculator to reduce the outside dependencies on this library and provide greater flexibility for client code (i.e. determining precision). Split out the validation methods and made them public (easier testability and it won't hurt to let the clients re-use this logic for their own validation).
Constructed a public enum for internally implemented formulas. Protected the formula strategies to internal use. Added a line in assembly info to expose these internal classes to the test project (when assembled under debug).

Created a constructor on the calculator class to take a client provided formula. Created another constructor to allow a client to select an internally implemented formula via the enum. This second constructor makes use of an internal factory to select the correct strategy.

Updated the demo to use an enum option to create the demo calculator.

Updated the tests to reflect the latest design. Removed the dependency on the calculator from the strategy tests. Implemented Math.Round in these tests to more reliably test the floating point values. The new factory needs some tests as well as the new constructor functionality in the calculator.