Extracted the valdation logic from the calculator and replaced it with an injectable abstraction. This validation is now configurable by the client. However, the previous validation code was put into a concrete class for validation and supplied by a new ValidatorFactory when needed.
Updated the tests as approrpiate, which mostly entailed removing more unnecessary dependencies. The validation logic tests no longer require the test harness to provide a formula, when a formula isn't necessary to validate the input.

The calculator received an upgrade to multiple constructors, supply default validator and formula where necessary. Along with the new validator factory above, the formula factory was expanded to provide a default formula to accomodate this and provide flexibility. It is anticipated that not every user will want to worry about what formula to pick or how to validate the input.

More unit tests surrounding the factories and the calculator are required.